azure-iot/azure-iot-read/requirements.txt
azure-iot-device
azure-iot-hub
azure-eventhub
asyncio
sh
